+++ src/routes/profile/+page.svelte
@@ -167,8 +167,14 @@
 						{/if}
 					{/if}
 					<tr>
-						<th>Project dir</th>
-						<td>{user.project_dir || '-'}</td>
+						<th>Project dirs</th>
+						<td>
+							<ul class="ps-3">
+								{#each user.project_dirs as dir (dir)}
+									<li>{dir}</li>
+								{/each}
+							</ul>
+						</td>
 					</tr>
 
 					{#if runnerBackend !== 'local'}