+++ src/routes/v2/admin/groups/+page.svelte
@@ -89,7 +89,6 @@
 				<th>Id</th>
 				<th>Name</th>
 				<th>Number of users</th>
-				<th>Number of viewer paths</th>
 				<th>Actions</th>
 			</tr>
 		</thead>
@@ -99,7 +98,6 @@
 					<td>{group.id}</td>
 					<td>{group.name}</td>
 					<td>{group.user_ids.length}</td>
-					<td>{group.viewer_paths.length}</td>
 					<td>
 						<a href="/v2/admin/groups/{group.id}" class="btn btn-light">
 							<i class="bi-info-circle"></i> Info